Issue #5 chore: Worked on PR Comments
