+++ src/components/AudioRecorderTamil/AudioRecorderTamil.js
@@ -9,7 +9,7 @@ import mic from '../../assests/Images/mic.png';
 import stop from '../../assests/Images/stop.png';
 import { showLoading, stopLoading } from '../../utils/Helper/SpinnerHandle';
 import { response,interact } from '../../services/telementryService';
-import { replaceAll } from '../../utils/helper';
+import { replaceAll, compareArrays } from '../../utils/helper';
 
 //webkitURL is deprecated but nevertheless
 URL = window.URL || window.webkitURL;
@@ -132,33 +132,6 @@ function Mic({
       ],
     });
 
-    function compareArrays(arr1, arr2) {
-        let words_result = [];
-
-        // Iterate over each element and compare
-        for (var k in arr1){
-            if (arr2[k] == "" || arr2[k] == undefined)
-            {
-                // Element not available on the same key
-                words_result.push('-1');
-            }
-            else if (arr1[k] === arr2[k]) {
-                // Elements match on the same key
-                words_result.push('1');
-            }
-            else {
-                // Element does not match on the same key
-                words_result.push('0');
-            }
-        }
-        if (arr1.length < arr2.length)
-        {
-           for (let i = arr1.length; i < arr2.length; i++){
-                words_result.push('-1');
-           }
-        }
-        return words_result;
-    }
     var requestOptions = {
       method: 'POST',
       headers: myHeaders,
